refactor pointer casting for cc warnings

+++ kernel/arch/x86/kernel/head64.c
@@ -30,7 +30,7 @@ static void syscall_init(void)
 {
     extern void syscall_entry();
     write_msr(IA32_STAR, 0, __KERNEL_CS | (__USER_CS32 << 16));
-    write_msr(IA32_LSTAR, (u32)syscall_entry, (u32)((uintptr_t)syscall_entry >> 32));
+    write_msr(IA32_LSTAR, (u32)(uintptr_t)syscall_entry, (u32)((uintptr_t)syscall_entry >> 32));
     write_msr(IA32_FMASK,
         X86_FLAGS_CF|X86_FLAGS_PF|X86_FLAGS_AF|
         X86_FLAGS_ZF|X86_FLAGS_SF|X86_FLAGS_TF|
